requirement.txt
streamlit==1.30.0
numpy==1.26.3
Pillow==10.2.0
opencv-python-headless==4.9.0.80
scikit-learn==1.4.0
